[#6] refactor: 회원 관련(레포지토리) 추상화
