+++ module-core/src/main/java/com/soundie/member/service/MemberService.java
@@ -2,7 +2,7 @@ package com.soundie.member.service;
 
 import com.soundie.member.domain.Member;
 import com.soundie.member.dto.MemberIdElement;
-import com.soundie.member.repository.MemberRepository;
+import com.soundie.member.repository.MemoryMemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class MemberService {
 
-    private final MemberRepository memberRepository;
+    private final MemoryMemberRepository memberRepository;
 
     public MemberIdElement createMember() {
         Member member = new Member("정원석");
