+++ module-core/src/main/java/com/soundie/post/service/PostService.java
@@ -3,7 +3,7 @@ package com.soundie.post.service;
 import com.soundie.global.common.exception.ApplicationError;
 import com.soundie.global.common.exception.NotFoundException;
 import com.soundie.member.domain.Member;
-import com.soundie.member.repository.MemberRepository;
+import com.soundie.member.repository.MemoryMemberRepository;
 import com.soundie.post.domain.Post;
 import com.soundie.post.domain.PostLike;
 import com.soundie.post.dto.*;
@@ -21,7 +21,7 @@ public class PostService {
 
     private final PostRepository postRepository;
     private final PostLikeRepository postLikeRepository;
-    private final MemberRepository memberRepository;
+    private final MemoryMemberRepository memberRepository;
 
     public GetPostResDto readPostList(){
         List<Post> findPosts = postRepository.findPosts();
